Implement BatchNorm with BNRS in the library
