+++ learn2learn/vision/models/__init__.py
@@ -31,8 +31,17 @@ from .cnn4 import (
     CNN4Backbone,
 )
 
+from .cnn4_bnrs import (
+    LinearBlock_BNRS,
+    ConvBlock_BNRS,
+    ConvBase_BNRS,
+    CNN4Backbone_BNRS,
+    CNN4_BNRS,
+)
+
 from .resnet12 import ResNet12, ResNet12Backbone
 from .wrn28 import WRN28, WRN28Backbone
+from .bnrs import BatchNorm_BNRS
 
 __all__ = [
     'get_pretrained_backbone',
@@ -49,6 +58,12 @@ __all__ = [
     'ResNet12Backbone',
     'WRN28',
     'WRN28Backbone',
+    'BatchNorm_BNRS',
+    'LinearBlock_BNRS',
+    'ConvBlock_BNRS',
+    'ConvBase_BNRS',
+    'CNN4Backbone_BNRS',
+    'CNN4_BNRS',
 ]
 
 _BACKBONE_URLS = {
